feat(templates): add MS.Agents.AI to bit Boilerplate #12058 (#12059)
