+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Features/Attachments/AttachmentController.cs
@@ -215,6 +215,30 @@ public partial class AttachmentController : AppControllerBase, IAttachmentContro
                 //#if (signalR == true || database == "PostgreSQL" || database == "SqlServer")
                 if (serviceProvider.GetService<IChatClient>() is IChatClient chatClient)
                 {
+                    var imageAnalysisAgent = chatClient.AsAIAgent(
+                        instructions: """
+                        You are a Product Image Specialist Agent. Your role is to analyze product images for an e-commerce catalog.
+
+                        ANALYSIS PROCESS:
+                        1. First, examine the image contents carefully
+                        2. Determine if the primary subject is a car (vehicle)
+                        3. If it is a car, provide a detailed, SEO-friendly description
+                        4. If it is NOT a car, explain why it doesn't meet catalog requirements
+
+                        RESPONSE FORMAT:
+                        Return ONLY a JSON object with:
+                        - "isCar": boolean (true if image shows a car, false otherwise)
+                        - "confidence": number between 0-1 indicating certainty of classification
+                        - "alt": string with detailed description for accessibility and SEO
+                        - "reasoning": string briefly explaining your analysis decision
+
+                        VALIDATION RULES:
+                        - Image quality must be acceptable for catalog use
+                        - Car must be clearly visible as the main subject
+                        """,
+                        name: "ProductImageAnalystAgent",
+                        description: "Analyzes product images to ensure they meet catalog standards for car products");
+
                     ChatOptions chatOptions = new()
                     {
                         ResponseFormat = ChatResponseFormat.Json,
@@ -226,21 +250,33 @@ public partial class AttachmentController : AppControllerBase, IAttachmentContro
 
                     configuration.GetRequiredSection("AI:ChatOptions").Bind(chatOptions);
 
-                    var response = await chatClient.GetResponseAsync<AIImageReviewResponse>(
+                    var response = await imageAnalysisAgent.RunAsync<AIImageReviewResponse>(
                         messages: [
-                            new ChatMessage(ChatRole.System, "Return a JSON object with two properties: 'isCar' (boolean, true if the image contains a car, false otherwise) and 'alt' (string, describing the image content, focusing on the car). Do not include any other properties or text."),
-                            new ChatMessage(ChatRole.User, "Analyze this image.")
+                            new ChatMessage(ChatRole.User, 
+                                "Analyze this product image for our car catalog. Is this a valid car product image that meets our quality and content standards?")
                             {
                                 Contents = [new DataContent(imageBytes, "image/webp")]
                             }
                         ],
                         cancellationToken: cancellationToken,
-                        options: chatOptions);
+                        options: new Microsoft.Agents.AI.ChatClientAgentRunOptions(chatOptions));
 
                     if (response.Result.IsCar is false)
                     {
-                        logger.LogWarning("Unexpected AI response for Car detection: {Response}", response.Result.Alt);
+                        logger.LogWarning(
+                            "Image validation failed - Not a car product. Confidence: {Confidence}, Reasoning: {Reasoning}", 
+                            response.Result.Confidence, 
+                            response.Result.Reasoning);
                         return BadRequest(Localizer[nameof(AppStrings.ImageNotCarError)].ToString());
+                    }
+
+                    if (response.Result.Confidence < 0.85)
+                    {
+                        logger.LogWarning(
+                            "Image analysis low confidence ({Confidence}). Reasoning: {Reasoning}. Alt text: {AltText}", 
+                            response.Result.Confidence,
+                            response.Result.Reasoning,
+                            response.Result.Alt);
                     }
 
                     altText = response.Result.Alt;
@@ -286,6 +322,6 @@ public partial class AttachmentController : AppControllerBase, IAttachmentContro
     }
 
     //#if (signalR == true || database == "PostgreSQL" || database == "SqlServer")
-    public record AIImageReviewResponse(bool IsCar, string? Alt);
+    public record AIImageReviewResponse(bool IsCar, double Confidence, string? Alt, string? Reasoning);
     //#endif
 }
