+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Infrastructure/SignalR/AppChatbot.cs
@@ -1,18 +1,8 @@
 ﻿//+:cnd:noEmit
 using System.Text;
-using System.ComponentModel;
 using System.Threading.Channels;
 using Boilerplate.Shared.Features.Chatbot;
-//#if (module == "Sales")
-using Boilerplate.Server.Api.Features.Products;
-//#endif
-using Boilerplate.Shared.Features.Identity.Dtos;
-using Microsoft.AspNetCore.SignalR;
 using Microsoft.AspNetCore.Authentication.BearerToken;
-using ModelContextProtocol.Server;
-using Boilerplate.Shared.Features.Diagnostic;
-using Boilerplate.Server.Api.Features.Identity;
-using Boilerplate.Server.Api.Infrastructure.Services;
 
 namespace Boilerplate.Server.Api.Infrastructure.SignalR;
 
@@ -20,9 +10,14 @@ namespace Boilerplate.Server.Api.Infrastructure.SignalR;
 /// Service responsible for managing chatbot conversations, maintaining chat history,
 /// and handling AI interactions including getting user feedbacks, describing app's features and pages etc.
 /// This service is exposed over SignalR's AppHub.Chat.cs, so it can accept stream of user messages and return stream of AI responses using AiChatPanel.razor
-/// Every tool method is decorated with [McpServerTool] attribute, so it can be also be used by other external MCP-Client if needed.
+/// Every tool method is decorated with [McpServerTool] attribute, so it can be also be used by other external MCP-Client if needed (Checkout AppChatbot.Tools.cs)
+/// 
+/// Microsoft.Agents.AI:
+/// Workflows are not implemented in this project, but with AIAgent, achieving them is now easier compared to using IChatClient directly.
+/// For example, it would be better to have separate Agents: one for product search, one for support, and one for app guidance.
+/// A coordinator Agent could determine which specialized Agent to delegate the task to based on the user's message,
+/// rather than having a single Agent with a very long System Prompt and many Tools.
 /// </summary>
-[McpServerToolType]
 public partial class AppChatbot
 {
     private IChatClient? chatClient = default!;
@@ -110,6 +105,11 @@ public partial class AppChatbot
 
         chatClient ??= serviceProvider.GetRequiredService<IChatClient>();
 
+        var supportAgent = chatClient.AsAIAgent(
+            instructions: supportSystemPrompt,
+            name: "SupportAgent",
+            description: "Provides user support, answers questions, and assists with app features and troubleshooting");
+
         StringBuilder assistantResponse = new();
         try
         {
@@ -119,6 +119,9 @@ public partial class AppChatbot
 
             // The following variables might change without SignalR connection restarts, so these should set here every time a new message is about to be processed.
             // For example, user can sign-in/sign-out during chat without restarting the app or SignalR connection.
+            // User can change these values using prompt injection, so it's important not to rely on these variables for critical logic or security decisions,
+            // but rather use them for providing better context to the model to generate more relevant responses.
+            // You can either check if user is authenticated or not at a time tools are being called, or add dedicated tool so the LLM model would call it to figure out the user's authentication state instead of relying on variables.
             var variablesPrompt = @$"
 ### Variables:
 {variablesDefault}
@@ -126,12 +129,10 @@ public partial class AppChatbot
 {{{{UserEmail}}}}: ""{(user.IsAuthenticated() ? user!.GetEmail()?.ToString() : "null")}""
 ";
 
-            await foreach (var response in chatClient.GetStreamingResponseAsync([
+            await foreach (var response in supportAgent.RunStreamingAsync([
                 new (ChatRole.System, variablesPrompt),
-                new (ChatRole.System, supportSystemPrompt),
-                    .. chatMessages,
-                    new (ChatRole.User, incomingMessage)
-                ], options: chatOptions, cancellationToken: cancellationToken))
+                .. chatMessages,
+                ], options: new Microsoft.Agents.AI.ChatClientAgentRunOptions(chatOptions), cancellationToken: cancellationToken))
             {
                 if (cancellationToken.IsCancellationRequested)
                     break;
@@ -194,274 +195,6 @@ public partial class AppChatbot
     }
 
     /// <summary>
-    /// Returns the current date and time based on the user's timezone.
-    /// </summary>
-    [Description("Returns the current date and time based on the user's timezone.")]
-    [McpServerTool(Name = nameof(GetCurrentDateTime))]
-    private string GetCurrentDateTime([Required, Description("User's timezone id")] string timeZoneId)
-    {
-        try
-        {
-            var timeZone = TimeZoneInfo.FindSystemTimeZoneById(timeZoneId);
-
-            var userDateTime = TimeZoneInfo.ConvertTime(DateTimeOffset.UtcNow, timeZone);
-
-            return $"Current date/time in user's timezone ({timeZoneId}) is {userDateTime:o}";
-        }
-        catch
-        {
-            return $"Current date/time in utc is {DateTimeOffset.UtcNow:o}";
-        }
-    }
-
-    /// <summary>
-    /// Saves the user's email address and the conversation history for future reference.
-    /// </summary>
-    [Description("Saves the user's email address and the conversation history for future reference.")]
-    [McpServerTool(Name = nameof(SaveUserEmailAndConversationHistory))]
-    private async Task<string?> SaveUserEmailAndConversationHistory(
-        [Required, Description("User's email address")] string emailAddress,
-        [Required, Description("Full conversation history")] string conversationHistory)
-    {
-        try
-        {
-            await using var scope = serviceProvider.CreateAsyncScope();
-
-            // Ideally, store these in a CRM or app database,
-            // but for now, we'll log them!
-            scope.ServiceProvider.GetRequiredService<ILogger<IChatClient>>()
-                .LogError("Chat reported issue: User email: {emailAddress}, Conversation history: {conversationHistory}", emailAddress, conversationHistory);
-
-            return "User email and conversation history saved successfully.";
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Failed to save user email and conversation history.";
-        }
-    }
-
-    /// <summary>
-    /// Navigates the user to a specific page within the application.
-    /// </summary>
-    [Description("Navigates the user to a specific page within the application. Use this tool when the user requests to go to a particular section or feature of the app.")]
-    [McpServerTool(Name = nameof(NavigateToPage))]
-    private async Task<string?> NavigateToPage(
-        [Required, Description("Page URL to navigate to")] string pageUrl)
-    {
-        await EnsureSignalRConnectionIdIsPresent();
-
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            _ = await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<bool>(SharedAppMessages.NAVIGATE_TO, pageUrl, CancellationToken.None);
-
-            return "Navigation completed";
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Navigation failed";
-        }
-    }
-
-    [Description(@"Displays the sign-in modal to the user and waits for either successful sign-in or cancellation")]
-    [McpServerTool(Name = nameof(ShowSignInModal))]
-    public async Task<UserDto?> ShowSignInModal()
-    {
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            await EnsureSignalRConnectionIdIsPresent();
-
-            var accessToken = await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<string>(SharedAppMessages.SHOW_SIGN_IN_MODAL, CancellationToken.None);
-
-            var bearerTokenProtector = bearerTokenOptions.Get(IdentityConstants.BearerScheme).BearerTokenProtector;
-            var accessTokenTicket = bearerTokenProtector.Unprotect(accessToken);
-            var user = accessTokenTicket!.Principal;
-
-            return await scope.ServiceProvider.GetRequiredService<AppDbContext>()
-                .Users
-                .Project()
-                .FirstOrDefaultAsync(u => u.Id == user.GetUserId());
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return null;
-        }
-    }
-
-    /// <summary>
-    /// Changes the user's culture/language setting.
-    /// </summary>
-    [Description("Changes the user's culture/language setting. Use this tool when the user requests to change the app language. Common LCIDs: 1033=en-US, 1065=fa-IR, 1053=sv-SE, 2057=en-GB, 1043=nl-NL, 1081=hi-IN, 2052=zh-CN, 3082=es-ES, 1036=fr-FR, 1025=ar-SA, 1031=de-DE.")]
-    [McpServerTool(Name = nameof(SetCulture))]
-    private async Task<string?> SetCulture(
-        [Required, Description("Culture LCID (e.g., 1033 for en-US, 1065 for fa-IR)")] int cultureLcid)
-    {
-        await EnsureSignalRConnectionIdIsPresent();
-
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            var culture = CultureInfo.GetCultureInfo(cultureLcid);
-
-            if (CultureInfoManager.SupportedCultures.All(c => c.Culture.LCID != cultureLcid))
-                return $"The requested culture is not supported. Available cultures: {string.Join(", ", CultureInfoManager.SupportedCultures.Select(c => c.Culture.NativeName))}";
-
-            _ = await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<bool>(SharedAppMessages.CHANGE_CULTURE, cultureLcid, CancellationToken.None);
-
-            return "Culture/Language changed successfully";
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Failed to change culture/language";
-        }
-    }
-
-    /// <summary>
-    /// Changes the user's theme preference between light and dark mode.
-    /// </summary>
-    [Description("Changes the user's theme preference between light and dark mode. Use this tool when the user requests to change the app theme or appearance.")]
-    [McpServerTool(Name = nameof(SetTheme))]
-    private async Task<string?> SetTheme(
-        [Required, Description("Theme name: 'light' or 'dark'")] string theme)
-    {
-        await EnsureSignalRConnectionIdIsPresent();
-
-        if (theme != "light" && theme != "dark")
-            return "Invalid theme. Use 'light' or 'dark'.";
-
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            _ = await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<bool>(SharedAppMessages.CHANGE_THEME, theme, CancellationToken.None);
-
-            return $"Theme changed to {theme} successfully";
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Failed to change theme";
-        }
-    }
-
-    /// <summary>
-    /// Retrieves the last error that occurred on the user's device from the diagnostic logs.
-    /// </summary>
-    [Description("Retrieves the last error that occurred on the user's device from the diagnostic logs. Use this tool when troubleshooting user-reported issues, investigating application crashes, or when the user mentions something isn't working.")]
-    [McpServerTool(Name = nameof(CheckLastError))]
-    private async Task<string?> CheckLastError()
-    {
-        await EnsureSignalRConnectionIdIsPresent();
-
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            var lastError = await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<DiagnosticLogDto?>(SharedAppMessages.UPLOAD_LAST_ERROR, CancellationToken.None);
-
-            if (lastError is null)
-                return "No errors found in the diagnostic logs.";
-
-            return lastError.ToString();
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Failed to retrieve error information from the device.";
-        }
-    }
-
-    /// <summary>
-    /// Clears application files on the user's device to fix issues.
-    /// </summary>
-    [Description("Clears application files on the user's device to fix issues.")]
-    [McpServerTool(Name = nameof(ClearAppFiles))]
-    private async Task<string?> ClearAppFiles()
-    {
-        await EnsureSignalRConnectionIdIsPresent();
-
-        await using var scope = serviceProvider.CreateAsyncScope();
-
-        try
-        {
-            await scope.ServiceProvider.GetRequiredService<IHubContext<AppHub>>()
-                .Clients.Client(signalRConnectionId!)
-                .InvokeAsync<DiagnosticLogDto?>(SharedAppMessages.CLEAR_APP_FILES, CancellationToken.None);
-
-            return "App files cleared successfully on the device.";
-        }
-        catch (Exception exp)
-        {
-            serviceProvider.GetRequiredService<ServerExceptionHandler>().Handle(exp);
-            return "Failed to clear app files on the device.";
-        }
-    }
-
-    //#if (module == "Sales")
-    //#if (database == "PostgreSQL" || database == "SqlServer")
-    /// <summary>
-    /// Searches for and recommends products based on user's needs and preferences.
-    /// </summary>
-    [Description("This tool searches for and recommends products based on a detailed description of the user's needs and preferences and returns recommended products.")]
-    [McpServerTool(Name = nameof(GetProductRecommendations))]
-    private async Task<object?> GetProductRecommendations(
-        [Required, Description("Concise summary of user requirements")] string userNeeds,
-        [Description("Car manufacturer's name (Optional)")] string? manufacturer,
-        [Description("Car price below this value (Optional)")] decimal? maxPrice,
-        [Description("Car price above this value (Optional)")] decimal? minPrice)
-    {
-        await using var scope = serviceProvider.CreateAsyncScope();
-        var productEmbeddingService = scope.ServiceProvider.GetRequiredService<ProductEmbeddingService>();
-        var searchQuery = string.IsNullOrWhiteSpace(manufacturer)
-            ? userNeeds
-            : $"**{manufacturer}** {userNeeds}";
-
-        // Get serverApiAddress from current context
-        var httpContextAccessor = scope.ServiceProvider.GetService<IHttpContextAccessor>();
-        var context = httpContextAccessor?.HttpContext;
-        var request = context?.Request;
-        Uri? serverApiAddress = request?.GetBaseUrl();
-
-        var recommendedProducts = await (await productEmbeddingService.SearchProducts(searchQuery, context?.RequestAborted ?? default))
-            .WhereIf(maxPrice.HasValue, p => p.Price <= maxPrice!.Value)
-            .WhereIf(minPrice.HasValue, p => p.Price >= minPrice!.Value)
-            .Take(10)
-            .Project()
-            .Select(p => new
-            {
-                p.Name,
-                p.PageUrl,
-                Manufacturer = p.CategoryName,
-                Price = p.FormattedPrice,
-                Description = p.DescriptionText,
-                PreviewImageUrl = p.GetPrimaryMediumImageUrl(serverApiAddress!) ?? "_content/Boilerplate.Client.Core/images/car_placeholder.png"
-            })
-            .ToArrayAsync(context?.RequestAborted ?? default);
-
-        return recommendedProducts;
-    }
-    //#endif
-    //#endif
-
-    /// <summary>
     /// Generate follow-up suggestions based on the conversation
     /// </summary>
     private async Task<AiChatFollowUpList> GenerateFollowUpSuggestions(
@@ -473,25 +206,45 @@ public partial class AppChatbot
         // This would generate a list of follow-up questions/suggestions to keep the conversation going.
         // You could instead generate that list in previous chat completion call:
         // 1: Using "tools" or "functions" feature of the model, that would not consider the latest assistant response.
-        // 2: Returning a json object containing the response and follow-up suggestions all together, losing IAsyncEnumerable streaming capability.  
+        // 2: Returning a json object containing the response and follow-up suggestions all together, losing IAsyncEnumerable streaming capability.
+
+        var followUpAgent = chatClient!.AsAIAgent(
+            instructions: """
+            You are a Follow-Up Suggestion Agent. Your role is to generate natural, contextual follow-up questions or actions for users.
+
+            ANALYSIS PROCESS:
+            1. Review the conversation context carefully
+            2. Identify logical next steps or questions the user might ask
+            3. Ensure suggestions are within the assistant's capabilities
+            4. Make suggestions actionable and user-centric
+
+            RESPONSE FORMAT:
+            Return ONLY a JSON object with:
+            - "FollowUpSuggestions": array of exactly 3 strings
+
+            VALIDATION RULES:
+            - Only suggest follow-up actions/questions that are within the assistant's scope and knowledge
+            - Do not suggest questions that require access to data or functionality that is unavailable or out of scope
+            - Avoid suggesting questions that the assistant would not be able to answer
+            - Written from the user's perspective (never from the assistant)
+            - Direct, natural, clickable actions/questions
+            - Keep each suggestion concise (under 60 characters)
+            """,
+            name: "FollowUpSuggestionAgent",
+            description: "Generates contextual follow-up suggestions to keep conversations flowing naturally");
+
         chatOptions.ResponseFormat = ChatResponseFormat.Json;
         chatOptions.AdditionalProperties = new() { ["response_format"] = new { type = "json_object" } };
 
-        var followUpItems = await chatClient!.GetResponseAsync<AiChatFollowUpList>([
-            new(ChatRole.System, supportSystemPrompt),
-            new(ChatRole.User, incomingMessage),
-            new(ChatRole.Assistant, assistantResponse),
-            new(ChatRole.System, """
-        Generate exactly 3 short suggested replies that the user would naturally type next.
-        Rules:
-        - Only suggest follow-up actions/questions that are within the assistant's scope and knowledge.
-        - Do not suggest questions that require access to data or functionality that is unavailable or out of scope for this assistant.
-        - Avoid suggesting questions that the assistant would not be able to answer.
-        - Written from the user's perspective (never from the assistant)
-        - Direct, natural, clickable actions/questions
-        - Return JSON object containing string[] named FollowUpSuggestions
-        """),],
-            chatOptions, cancellationToken: cancellationToken);
+        var followUpItems = await followUpAgent.RunAsync<AiChatFollowUpList>(
+            messages: [
+                new(ChatRole.System, supportSystemPrompt),
+                new(ChatRole.User, incomingMessage),
+                new(ChatRole.Assistant, assistantResponse),
+                new(ChatRole.User, "Generate 3 short follow-up suggestions for what I might want to ask or do next.")
+            ],
+            cancellationToken: cancellationToken,
+            options: new Microsoft.Agents.AI.ChatClientAgentRunOptions(chatOptions));
 
         return followUpItems.Result ?? new AiChatFollowUpList();
     }
